rename files to proto_value
